Split unstructured.go into several parts
No functional changes
