+++ staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured/unstructured_list_test.go
@@ -17,30 +17,8 @@ limitations under the License.
 package unstructured
 
 import (
-	"io/ioutil"
-	"sync"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
-
-// TestCodecOfUnstructuredList tests that there are no data races in Encode().
-// i.e. that it does not mutate the object being encoded.
-func TestCodecOfUnstructuredList(t *testing.T) {
-	var wg sync.WaitGroup
-	concurrency := 10
-	list := UnstructuredList{
-		Object: map[string]interface{}{},
-	}
-	wg.Add(concurrency)
-	for i := 0; i < concurrency; i++ {
-		go func() {
-			defer wg.Done()
-			assert.NoError(t, UnstructuredJSONScheme.Encode(&list, ioutil.Discard))
-		}()
-	}
-	wg.Wait()
-}
 
 func TestUnstructuredList(t *testing.T) {
 	list := &UnstructuredList{
